Extract Ktor plugin into module
