+++ networklogger/ktor/src/main/java/dev/kgbier/util/networklogger/ktor/Public.kt
@@ -1,0 +1,10 @@
+package dev.kgbier.util.networklogger.ktor
+
+import android.content.Context
+import dev.kgbier.util.networklogger.NetworkLoggerRepository
+import io.ktor.client.plugins.*
+
+typealias NetworkLoggerKtorPluginInstaller = HttpClientPlugin<Nothing, NetworkLoggerKtorPlugin>
+
+fun NetworkLoggerKtorPluginInstaller(context: Context): NetworkLoggerKtorPluginInstaller =
+    NetworkLoggerKtorPlugin.NetworkLoggerKtorPluginInstaller(NetworkLoggerRepository(context))
