+++ sample/build.gradle
@@ -27,13 +27,14 @@ android.variantFilter { variant ->
 dependencies {
 
     implementation project(path: ':networklogger:okhttp')
+    implementation project(path: ':networklogger:ktor')
 
     implementation 'com.squareup.okhttp3:okhttp:4.10.0'
     implementation 'com.squareup.okhttp3:logging-interceptor:4.10.0'
 
     implementation 'io.ktor:ktor-client-core:2.0.3'
-    implementation("io.ktor:ktor-client-cio:2.0.3")
-    implementation("io.ktor:ktor-client-logging:2.0.3")
+    implementation 'io.ktor:ktor-client-cio:2.0.3'
+    implementation 'io.ktor:ktor-client-logging:2.0.3'
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.4.2'
